int64 indexing: comprehensive migration progress
Core Types:
- Shape.cs, Shape.Reshaping.cs: int to long for dimensions, strides, offsets
- ArraySlice.cs, ArraySlice`1.cs: int64 indexing support
- UnmanagedStorage.cs, Getters, Setters: long indices throughout
- UnmanagedMemoryBlock`1.cs, Casting.cs: memory block int64 support
- Slice.cs: slice indices as long

Iterators:
- NDIterator.cs, MultiIterator.cs: long coordinate support
- All Incrementors: NDCoordinatesIncrementor, AxisIncrementor, etc.

IL Kernels:
- ILKernelGenerator (Binary, Unary, Comparison, Reduction, Scan, Shift, Clip, Modf, Masking): long offsets
- KernelSignatures.cs, BinaryKernel.cs, ReductionKernel.cs: delegate signatures updated
- SimdReductionOptimized.cs, SimdThresholds.cs: SIMD path updates

DefaultEngine:
- All reduction operations (Add, AMax, AMin, ArgMax, ArgMin, Mean, Std, Var, etc.)
- BinaryOp, UnaryOp, CompareOp: long indices
- Dot.NDMD, MatMul.2D2D, Clip, Modf, Shift: BLAS operations updated

NDArray & APIs:
- NDArray.cs, NDArray.Unmanaged.cs, NDArray.String.cs: core array updates
- TensorEngine.cs, NDArray`1.cs: engine interface updates
- Indexing (Selection, Masking): long index support
- np.size, np.array, np.zeros, np.ones, np.full, np.meshgrid, etc.

New files:
- np.count_nonzero.cs: new API implementation
- Default.IsInf.cs: infinity checking
- CountNonzeroTests.cs: test coverage

Tests updated across all affected components.

Part of int64 indexing migration (#584)

+++ src/NumSharp.Core/Backends/Unmanaged/ArraySlice`1.cs
@@ -137,7 +137,7 @@ namespace NumSharp.Backends.Unmanaged
         }
 
         [MethodImpl(OptimizeAndInline)]
-        public void SetIndex(long index, T value)
+        public void SetIndex(int index, T value)
         {
             Debug.Assert(index < Count, "index < Count, Memory corruption expected.");
             *(Address + index) = value;
